refactor: Reorganize code architecture with modular structure
This commit refactors the chess application from large monolithic files
into a clean, modular architecture following Single Responsibility Principle.

Key Changes:
- Consolidated chess utilities from 440-line chessLogic.ts into focused modules:
  - gameBoard.ts (~225 lines): Piece utilities + game state detection
  - moveRules.ts (~340 lines): Move validation + execution

- Refactored React hooks from 357-line useChessGame.ts into composition:
  - useGameBoard.ts (~100 lines): Board and UI state management
  - useComputerAI.ts (~55 lines): AI orchestration
  - useMoveRules.ts (~270 lines): Move execution logic

- Added AI Easy mode with random legal move selection
  - utils/ai/easyAI.ts: AI move logic
  - Computer makes moves after 1-second thinking delay
  - "Computer is thinking..." UI feedback

- Cleaned up legacy files:
  - Removed BUGS.md (issues resolved)
  - Removed src/scripts/main.js (vanilla JS, no longer needed)
  - Removed empty scripts folder

Architecture Benefits:
- Clear separation of concerns (board state, move logic, AI, UI)
- Smaller, more maintainable files
- Consistent naming (gameBoard.ts ↔ useGameBoard, moveRules.ts ↔ useMoveRules)
- Easier testing and future enhancements
- Reduced code by ~2,700 lines through better organization

All chess features remain fully functional:
- Complete move validation (check, checkmate, stalemate)
- Special moves (castling, en passant, pawn promotion)
- Game modes (Friend vs Friend, AI Easy)
- Move history and captured pieces tracking
