+++ TODO.md
@@ -1,0 +1,1 @@
+- Add easy computer
