+++ src/utils/ai/index.ts
@@ -1,0 +1,4 @@
+// Re-export AI utilities
+
+export * from './easyAI'
+// Future: export * from './hardAI' (Stockfish integration)
