+++ src/utils/chess/index.ts
@@ -1,0 +1,4 @@
+// Re-export all chess logic functions from organized modules
+
+export * from './gameBoard'
+export * from './moveRules'
