+++ src/utils/chess/moveRules.ts
@@ -1,41 +1,19 @@
+/**
+ * Chess Move Validation Rules
+ * Contains all logic for validating if a move is legal
+ */
+
 import {
   BoardState,
-  PieceSymbol,
   PieceColor,
-  PieceType,
+  PieceSymbol,
   Position,
-  PIECES,
-  PIECE_VALUES,
-} from '../types/chess'
+} from '../../types/chess'
+import { getPieceInfo } from './gameBoard'
 
-export interface PieceInfo {
-  type: PieceType
-  color: PieceColor
-  symbol: PieceSymbol
-}
-
-export function getPieceInfo(piece: PieceSymbol): PieceInfo | null {
-  if (!piece) return null
-
-  const whiteValues = Object.values(PIECES.white)
-  const blackValues = Object.values(PIECES.black)
-
-  if (whiteValues.includes(piece)) {
-    const type = Object.keys(PIECES.white).find(
-      (key) => PIECES.white[key as PieceType] === piece
-    ) as PieceType
-    return { type, color: 'white', symbol: piece }
-  }
-
-  if (blackValues.includes(piece)) {
-    const type = Object.keys(PIECES.black).find(
-      (key) => PIECES.black[key as PieceType] === piece
-    ) as PieceType
-    return { type, color: 'black', symbol: piece }
-  }
-
-  return null
-}
+// ============================================================================
+// PIECE-SPECIFIC MOVEMENT RULES
+// ============================================================================
 
 export function isValidPawnMove(
   fromRow: number,
@@ -172,13 +150,17 @@ export function isValidKingMove(
     return true
   }
 
-  // Castling (2 squares horizontally) - further validation in makeMove
+  // Castling (2 squares horizontally) - further validation in isValidCastling
   if (rowDiff === 0 && colDiff === 2) {
     return true
   }
 
   return false
 }
+
+// ============================================================================
+// SPECIAL MOVE VALIDATION
+// ============================================================================
 
 export function isValidCastling(
   fromRow: number,
@@ -221,6 +203,10 @@ export function isValidCastling(
 
   return true
 }
+
+// ============================================================================
+// GENERAL MOVE VALIDATION
+// ============================================================================
 
 export function isValidMove(
   fromRow: number,
@@ -283,126 +269,71 @@ export function getValidMoves(
   return validMoves
 }
 
-export function getPieceValue(piece: PieceSymbol): number {
-  const pieceInfo = getPieceInfo(piece)
-  if (!pieceInfo) return 0
-  return PIECE_VALUES[pieceInfo.type]
+// ============================================================================
+// MOVE EXECUTION
+// ============================================================================
+
+export interface MoveExecutionResult {
+  newBoard: BoardState
+  capturedPiece: PieceSymbol | null
+  isCastling: boolean
+  isEnPassant: boolean
 }
 
-// Find the king position for a given color
-export function findKingPosition(boardState: BoardState, color: PieceColor): Position | null {
-  for (let row = 0; row < 8; row++) {
-    for (let col = 0; col < 8; col++) {
-      const piece = boardState[row][col]
-      if (piece) {
-        const pieceInfo = getPieceInfo(piece)
-        if (pieceInfo && pieceInfo.type === 'king' && pieceInfo.color === color) {
-          return { row, col }
-        }
-      }
-    }
-  }
-  return null
-}
-
-// Check if a square is under attack by the opponent
-export function isSquareUnderAttack(
-  row: number,
-  col: number,
-  attackerColor: PieceColor,
-  boardState: BoardState
-): boolean {
-  // Check all opponent pieces to see if they can attack this square
-  for (let fromRow = 0; fromRow < 8; fromRow++) {
-    for (let fromCol = 0; fromCol < 8; fromCol++) {
-      const piece = boardState[fromRow][fromCol]
-      if (piece) {
-        const pieceInfo = getPieceInfo(piece)
-        if (pieceInfo && pieceInfo.color === attackerColor) {
-          // For pawns, check diagonal attacks only (not forward moves)
-          if (pieceInfo.type === 'pawn') {
-            const direction = attackerColor === 'white' ? -1 : 1
-            const rowDiff = row - fromRow
-            const colDiff = Math.abs(col - fromCol)
-            if (rowDiff === direction && colDiff === 1) {
-              return true
-            }
-          } else {
-            // For other pieces, use normal move validation
-            if (isValidMove(fromRow, fromCol, row, col, boardState)) {
-              return true
-            }
-          }
-        }
-      }
-    }
-  }
-  return false
-}
-
-// Check if the king is in check
-export function isInCheck(color: PieceColor, boardState: BoardState): boolean {
-  const kingPos = findKingPosition(boardState, color)
-  if (!kingPos) return false
-
-  const opponentColor: PieceColor = color === 'white' ? 'black' : 'white'
-  return isSquareUnderAttack(kingPos.row, kingPos.col, opponentColor, boardState)
-}
-
-// Check if a move would leave the king in check
-export function wouldBeInCheck(
+/**
+ * Executes a chess move on the board (without validation)
+ * Handles special moves: castling, en passant
+ */
+export function executeMove(
   fromRow: number,
   fromCol: number,
   toRow: number,
   toCol: number,
-  boardState: BoardState
-): boolean {
+  boardState: BoardState,
+  enPassantTarget: Position | null,
+  isCastling: boolean
+): MoveExecutionResult {
   const piece = boardState[fromRow][fromCol]
-  if (!piece) return false
+  const pieceInfo = getPieceInfo(piece!)
 
-  const pieceInfo = getPieceInfo(piece)
-  if (!pieceInfo) return false
+  // Copy the board
+  const newBoard = boardState.map((row) => [...row])
+  let capturedPiece = newBoard[toRow][toCol]
 
-  // Simulate the move
-  const newBoard = boardState.map(row => [...row])
+  // Move the piece
   newBoard[toRow][toCol] = piece
   newBoard[fromRow][fromCol] = null
 
-  return isInCheck(pieceInfo.color, newBoard)
-}
-
-// Check if the current player has any legal moves
-export function hasLegalMoves(color: PieceColor, boardState: BoardState, enPassantTarget?: Position | null): boolean {
-  for (let fromRow = 0; fromRow < 8; fromRow++) {
-    for (let fromCol = 0; fromCol < 8; fromCol++) {
-      const piece = boardState[fromRow][fromCol]
-      if (piece) {
-        const pieceInfo = getPieceInfo(piece)
-        if (pieceInfo && pieceInfo.color === color) {
-          // Check all possible destination squares
-          for (let toRow = 0; toRow < 8; toRow++) {
-            for (let toCol = 0; toCol < 8; toCol++) {
-              if (isValidMove(fromRow, fromCol, toRow, toCol, boardState, enPassantTarget)) {
-                // Check if this move would leave the king in check
-                if (!wouldBeInCheck(fromRow, fromCol, toRow, toCol, boardState)) {
-                  return true
-                }
-              }
-            }
-          }
-        }
-      }
-    }
+  // Handle castling - move the rook too
+  if (isCastling) {
+    const isKingSide = toCol > fromCol
+    const rookFromCol = isKingSide ? 7 : 0
+    const rookToCol = isKingSide ? toCol - 1 : toCol + 1
+    const rook = newBoard[fromRow][rookFromCol]
+    newBoard[fromRow][rookToCol] = rook
+    newBoard[fromRow][rookFromCol] = null
   }
-  return false
-}
 
-// Check if it's checkmate
-export function isCheckmate(color: PieceColor, boardState: BoardState, enPassantTarget?: Position | null): boolean {
-  return isInCheck(color, boardState) && !hasLegalMoves(color, boardState, enPassantTarget)
-}
+  let isEnPassant = false
 
-// Check if it's stalemate
-export function isStalemate(color: PieceColor, boardState: BoardState, enPassantTarget?: Position | null): boolean {
-  return !isInCheck(color, boardState) && !hasLegalMoves(color, boardState, enPassantTarget)
+  // Handle en passant capture
+  if (
+    pieceInfo?.type === 'pawn' &&
+    enPassantTarget &&
+    toRow === enPassantTarget.row &&
+    toCol === enPassantTarget.col &&
+    !capturedPiece
+  ) {
+    const capturedPawnRow = pieceInfo.color === 'white' ? toRow + 1 : toRow - 1
+    capturedPiece = newBoard[capturedPawnRow][toCol]
+    newBoard[capturedPawnRow][toCol] = null
+    isEnPassant = true
+  }
+
+  return {
+    newBoard,
+    capturedPiece,
+    isCastling,
+    isEnPassant,
+  }
 }
